# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "simple description",
        "title": "QrMenuApi",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/company/{companyID}/pubs/{pubID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get pub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Get pub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pubs.getPubByIDOutput"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates pub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Update pub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "pub params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.UpdatePubInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pubs.updatePubOutput"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes pub",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Delete pub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pubs.deltePubOutput"
                        }
                    }
                }
            }
        },
        "/company/{companyID}/pubs/{pubID}/bg": {
            "get": {
                "description": "get pub bg",
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Get pub bg",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads pub bg(png)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Upload bg",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "bg",
                        "name": "bg",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pubs.fileUploadOutput"
                        }
                    }
                }
            }
        },
        "/company/{companyID}/pubs/{pubID}/logo": {
            "get": {
                "description": "get pub logo",
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Get pub logo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads pub logo(png)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Upload logo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "logo",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pubs.fileUploadOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.PubOutput": {
            "type": "object",
            "properties": {
                "additional_info": {
                    "type": "string",
                    "example": "My pub additional info"
                },
                "address": {
                    "type": "string",
                    "example": "My pub address"
                },
                "bg_image_file_name": {
                    "type": "string",
                    "example": "my-pub-name-bg.png"
                },
                "color": {
                    "type": "string",
                    "example": "#ffffff"
                },
                "color_theme": {
                    "type": "string",
                    "example": "light"
                },
                "company_id": {
                    "type": "integer"
                },
                "currency_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "logo_file_name": {
                    "type": "string",
                    "example": "my-pub-name-logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "My pub name"
                },
                "qr_code_file_name": {
                    "type": "string",
                    "example": "my-pub-name.png"
                },
                "url_name": {
                    "type": "string",
                    "example": "my-pub-name"
                },
                "wifi_password": {
                    "type": "string",
                    "example": "12345678"
                }
            }
        },
        "entities.UpdatePubInput": {
            "type": "object",
            "required": [
                "additional_info",
                "address",
                "color",
                "color_theme",
                "name",
                "url_name",
                "wifi_password"
            ],
            "properties": {
                "additional_info": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My pub additional info"
                },
                "address": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My pub address"
                },
                "color": {
                    "type": "string",
                    "minLength": 2,
                    "example": "#ffffff"
                },
                "color_theme": {
                    "type": "string",
                    "minLength": 2,
                    "example": "light"
                },
                "company_id": {
                    "type": "integer",
                    "example": 1
                },
                "currency_id": {
                    "type": "integer",
                    "example": 22
                },
                "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My pub name"
                },
                "url_name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "my-pub-name"
                },
                "wifi_password": {
                    "type": "string",
                    "minLength": 2,
                    "example": "12345678"
                }
            }
        },
...
        "pubs.deltePubOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "pubs.fileUploadOutput": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "pubs.getPubByIDOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "pub": {
                    "$ref": "#/definitions/entities.PubOutput"
                }
            }
        },
        "pubs.updatePubOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "pub": {
                    "$ref": "#/definitions/entities.PubOutput"
                }
            }
        }
    }
}